Timer system bones -BH
tick manager has the event for when a tick happens and how long a tick is, if there is a timer, and has the control for tickNow.

other classes can simply call on: TickManager.tick += TheirFunction; to get their function to be called on the coming tick's, and: TickManager.tick -= TheirFunction; to remove their function from the tick system.