Reworked the world transfer code -NB
basically just changed how the world transfer was handled using some examples found on the unity forums. added in 2 new scripts called player destroyer and camera destroyer that will destroy any instance of the main camera or the player that the script is attached to after the first instance is created. also changed the coroutines so that they allow for the spawn point game objects to be found and manipulated.