Added some enemies to the overworld, tweaked how entering combat with them works -NB
The enemies are just placed around the map in areas i think makes sense for them to be, could be moved around no problem. Entering combat will now move an empty game object to where the player was when the collision with the enemy occurred, can use that game object to place the player where they were in the world when they entered combat.