no real progress :(
made a showcase of the animals that are in the package were using, and looked at a few effects that might be worth using.

attempting to make a hose of ice that when touched unrendered blocks will make them revealed or re-rendered. Need to study for a calculus test for now though, this was what I was working off of, https://unity3d.com/learn/tutorials/lets-try/shooting-with-raycasts?playlist=41639