Updated the Ice World Platforming to be more doable -NB
ice world platforming is now able to be done, the ramp functions, it's not perfect but it works. The platforms at the max height can be dashed to with the fire demon's dash. The UI has been added to the build order as well, it is the default scene and hitting play loads into the hub world.