Added a ramp to the start of the ice world, added colliders on the fire tornado and fire walls. -NB
A ramp in the start of the ice world to make it easier to understand where to go, colliders added to the tornado and the wall in the fire world that shoots the player back to the fire world if they walk into fire. Might've broken the particle colliding with the fire to turn it off though.