Inventory canvas object and inventory manager moved to follow the player
Inventory canvas object and script now start in the hub world and move between worlds with the player, so the data is persistent. Added a script to destroy any extra inventory objects that get created, and a check in the inventory manager that'll find the tick manager script when we enter combat.